+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/traffic_signal_state.hpp
@@ -1,4 +1,4 @@
-// Copyright 2015-2020 Tier IV, Inc. All rights reserved.
+// Copyright 2015 TIER IV, Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -40,7 +40,7 @@ struct TrafficSignalState
    *  ID of the referenced signal in a road network. The signal ID must be
    *  listed in TrafficSignal list of the RoadNetwork.
    *
-   *  In the TierIV OpenSCENARIO implementation, it is the Lanelet ID (positive
+   *  In the TIER IV OpenSCENARIO implementation, it is the Lanelet ID (positive
    *  integer) of the traffic light.
    *
    * ------------------------------------------------------------------------ */