+++ openscenario/openscenario_interpreter/src/syntax/scenario_object.cpp
@@ -1,4 +1,4 @@
-// Copyright 2015-2021 Tier IV, Inc. All rights reserved.
+// Copyright 2015 TIER IV, Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -69,21 +69,38 @@ auto ScenarioObject::activateSensors() -> bool
      described in ScenarioObject.ObjectController.Controller.Properties as
      "characteristics of the person driving the car.
   */
-  simulation_api_schema::DetectionSensorConfiguration configuration;
+  simulation_api_schema::DetectionSensorConfiguration detection_sensor_configuration;
   {
-    configuration.set_entity(name);
-    configuration.set_architecture_type(
+    detection_sensor_configuration.set_entity(name);
+    detection_sensor_configuration.set_architecture_type(
       getParameter<std::string>("architecture_type", "awf/universe"));
-    configuration.set_update_duration(0.1);
-    configuration.set_range(300);
-    configuration.set_filter_by_range(
+    detection_sensor_configuration.set_update_duration(0.1);
+    detection_sensor_configuration.set_range(300);
+    detection_sensor_configuration.set_filter_by_range(
+      object_controller.is<Controller>()
+        ? object_controller.as<Controller>().properties.get<Boolean>("isClairvoyant")
+        : false);
+  }
+
+  simulation_api_schema::OccupancyGridSensorConfiguration occupancy_grid_sensor_configuration;
+  {
+    occupancy_grid_sensor_configuration.set_entity(name);
+    occupancy_grid_sensor_configuration.set_architecture_type(
+      getParameter<std::string>("architecture_type", "awf/universe"));
+    occupancy_grid_sensor_configuration.set_update_duration(0.1);
+    occupancy_grid_sensor_configuration.set_resolution(0.5);
+    occupancy_grid_sensor_configuration.set_width(200);
+    occupancy_grid_sensor_configuration.set_height(200);
+    occupancy_grid_sensor_configuration.set_range(300);
+    occupancy_grid_sensor_configuration.set_filter_by_range(
       object_controller.is<Controller>()
         ? object_controller.as<Controller>().properties.get<Boolean>("isClairvoyant")
         : false);
   }
 
   return object_controller.isUserDefinedController() and attachLidarSensor(name) and
-         attachDetectionSensor(configuration);
+         attachDetectionSensor(detection_sensor_configuration) and
+         attachOccupancyGridSensor(occupancy_grid_sensor_configuration);
 }
 }  // namespace syntax
 }  // namespace openscenario_interpreter