+++ simulation/simple_sensor_simulator/src/sensor_simulation/primitives/primitive.cpp
@@ -1,4 +1,4 @@
-// Copyright 2015-2020 Tier IV, Inc. All rights reserved.
+// Copyright 2015 TIER IV, Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,6 +15,9 @@
 #include <quaternion_operation/quaternion_operation.h>
 
 #include <algorithm>
+#include <boost/geometry.hpp>
+#include <boost/geometry/geometries/point_xy.hpp>
+#include <boost/geometry/geometries/polygon.hpp>
 #include <iostream>
 #include <simple_sensor_simulator/sensor_simulation/primitives/primitive.hpp>
 #include <string>
@@ -24,7 +27,7 @@ namespace simple_sensor_simulator
 {
 namespace primitives
 {
-Primitive::Primitive(std::string type, geometry_msgs::msg::Pose pose)
+Primitive::Primitive(std::string type, const geometry_msgs::msg::Pose & pose)
 : type(std::move(type)), pose(pose)
 {
 }
@@ -47,6 +50,25 @@ Vertex Primitive::transform(Vertex v) const
   return ret;
 }
 
+Vertex Primitive::transform(Vertex v, const geometry_msgs::msg::Pose & sensor_pose) const
+{
+  auto mat = quaternion_operation::getRotationMatrix(
+    quaternion_operation::getRotation(sensor_pose.orientation, pose.orientation));
+  Eigen::VectorXd point(3);
+  point(0) = v.x;
+  point(1) = v.y;
+  point(2) = v.z;
+  point = mat * point;
+  point(0) = point(0) + pose.position.x - sensor_pose.position.x;
+  point(1) = point(1) + pose.position.y - sensor_pose.position.y;
+  point(2) = point(2) + pose.position.z - sensor_pose.position.z;
+  Vertex ret;
+  ret.x = point(0);
+  ret.y = point(1);
+  ret.z = point(2);
+  return ret;
+}
+
 std::vector<Vertex> Primitive::transform() const
 {
   std::vector<Vertex> ret;
@@ -57,6 +79,73 @@ std::vector<Vertex> Primitive::transform() const
   return ret;
 }
 
+<<<<<<< HEAD
+=======
+std::vector<Vertex> Primitive::transform(const geometry_msgs::msg::Pose & sensor_pose) const
+{
+  std::vector<Vertex> ret;
+  for (auto & v : vertices_) {
+    ret.emplace_back(transform(v, sensor_pose));
+  }
+  return ret;
+}
+
+std::vector<Vertex> Primitive::getVertex() const { return transform(); }
+
+std::vector<Triangle> Primitive::getTriangles() const { return triangles_; }
+
+std::vector<geometry_msgs::msg::Point> Primitive::get2DConvexHull(
+  const geometry_msgs::msg::Pose & sensor_pose) const
+{
+  const auto vertex = transform(sensor_pose);
+  typedef boost::geometry::model::d2::point_xy<double> boost_point;
+  typedef boost::geometry::model::polygon<boost_point> boost_polygon;
+  boost_polygon poly;
+  for (const auto & p : vertex) {
+    boost::geometry::exterior_ring(poly).push_back(boost_point(p.x, p.y));
+  }
+  boost_polygon hull;
+  boost::geometry::convex_hull(poly, hull);
+  std::vector<geometry_msgs::msg::Point> polygon;
+  for (auto it = boost::begin(boost::geometry::exterior_ring(hull));
+       it != boost::end(boost::geometry::exterior_ring(hull)); ++it) {
+    double x = boost::geometry::get<0>(*it);
+    double y = boost::geometry::get<1>(*it);
+    geometry_msgs::msg::Point p;
+    p.x = x;
+    p.y = y;
+    p.z = 0.0;
+    polygon.emplace_back(p);
+  }
+  return polygon;
+}
+
+std::vector<geometry_msgs::msg::Point> Primitive::get2DConvexHull() const
+{
+  const auto vertex = getVertex();
+  typedef boost::geometry::model::d2::point_xy<double> boost_point;
+  typedef boost::geometry::model::polygon<boost_point> boost_polygon;
+  boost_polygon poly;
+  for (const auto & p : vertex) {
+    boost::geometry::exterior_ring(poly).push_back(boost_point(p.x, p.y));
+  }
+  boost_polygon hull;
+  boost::geometry::convex_hull(poly, hull);
+  std::vector<geometry_msgs::msg::Point> polygon;
+  for (auto it = boost::begin(boost::geometry::exterior_ring(hull));
+       it != boost::end(boost::geometry::exterior_ring(hull)); ++it) {
+    double x = boost::geometry::get<0>(*it);
+    double y = boost::geometry::get<1>(*it);
+    geometry_msgs::msg::Point p;
+    p.x = x;
+    p.y = y;
+    p.z = 0.0;
+    polygon.emplace_back(p);
+  }
+  return polygon;
+}
+
+>>>>>>> 6b642a14fb657505da431412d45b696e5bc9ace8
 unsigned int Primitive::addToScene(RTCDevice device, RTCScene scene)
 {
   RTCGeometry mesh = rtcNewGeometry(device, RTC_GEOMETRY_TYPE_TRIANGLE);
@@ -76,6 +165,100 @@ unsigned int Primitive::addToScene(RTCDevice device, RTCScene scene)
   unsigned int geometry_id = rtcAttachGeometry(scene, mesh);
   rtcReleaseGeometry(mesh);
   return geometry_id;
+}
+
+boost::optional<double> Primitive::getMax(const Axis & axis) const
+{
+  std::vector<double> values;
+  if (vertices_.empty()) {
+    return boost::none;
+  }
+  const auto transformed_vertices = transform();
+  for (const auto v : transformed_vertices) {
+    switch (axis) {
+      case Axis::X:
+        values.emplace_back(v.x);
+        break;
+      case Axis::Y:
+        values.emplace_back(v.y);
+        break;
+      case Axis::Z:
+        values.emplace_back(v.z);
+        break;
+    }
+  }
+  return *std::max_element(values.begin(), values.end());
+}
+
+boost::optional<double> Primitive::getMin(const Axis & axis) const
+{
+  std::vector<double> values;
+  if (vertices_.empty()) {
+    return boost::none;
+  }
+  const auto transformed_vertices = transform();
+  for (const auto v : transformed_vertices) {
+    switch (axis) {
+      case Axis::X:
+        values.emplace_back(v.x);
+        break;
+      case Axis::Y:
+        values.emplace_back(v.y);
+        break;
+      case Axis::Z:
+        values.emplace_back(v.z);
+        break;
+    }
+  }
+  return *std::min_element(values.begin(), values.end());
+}
+
+boost::optional<double> Primitive::getMax(
+  const Axis & axis, const geometry_msgs::msg::Pose & sensor_pose) const
+{
+  std::vector<double> values;
+  if (vertices_.empty()) {
+    return boost::none;
+  }
+  const auto transformed_vertices = transform(sensor_pose);
+  for (const auto v : transformed_vertices) {
+    switch (axis) {
+      case Axis::X:
+        values.emplace_back(v.x);
+        break;
+      case Axis::Y:
+        values.emplace_back(v.y);
+        break;
+      case Axis::Z:
+        values.emplace_back(v.z);
+        break;
+    }
+  }
+  return *std::max_element(values.begin(), values.end());
+}
+
+boost::optional<double> Primitive::getMin(
+  const Axis & axis, const geometry_msgs::msg::Pose & sensor_pose) const
+{
+  std::vector<double> values;
+  if (vertices_.empty()) {
+    return boost::none;
+  }
+  const auto transformed_vertices = transform(sensor_pose);
+  for (const auto v : transformed_vertices) {
+    switch (axis) {
+      case Axis::X:
+        values.emplace_back(v.x);
+        break;
+      case Axis::Y:
+        values.emplace_back(v.y);
+        break;
+      case Axis::Z:
+        values.emplace_back(v.z);
+        break;
+    }
+  }
+  return *std::min_element(values.begin(), values.end());
 }
 }  // namespace primitives
 }  // namespace simple_sensor_simulator