+++ simulation/traffic_simulator/include/traffic_simulator/entity/ego_entity.hpp
@@ -1,4 +1,4 @@
-// Copyright 2015-2020 Tier IV, Inc. All rights reserved.
+// Copyright 2015 TIER IV, Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -90,6 +90,8 @@ public:
 
   auto getDriverModel() const -> traffic_simulator_msgs::msg::DriverModel override;
 
+  auto getEmergencyStateName() const -> std::string override;
+
   auto getEntityStatus(const double, const double) const
     -> const traffic_simulator_msgs::msg::EntityStatus;
 
@@ -99,13 +101,13 @@ public:
 
   auto getRouteLanelets() const -> std::vector<std::int64_t>;
 
+  auto getTurnIndicatorsCommandName() const -> std::string override;
+
   auto getVehicleCommand() const -> std::tuple<
     autoware_auto_control_msgs::msg::AckermannControlCommand,
     autoware_auto_vehicle_msgs::msg::GearCommand> override;
 
   auto getWaypoints() -> const traffic_simulator_msgs::msg::WaypointsArray override;
-
-  auto getEmergencyStateString() const -> std::string override;
 
   void onUpdate(double current_time, double step_time) override;