+++ simulation/traffic_simulator/src/api/api.cpp
@@ -1,4 +1,4 @@
-// Copyright 2015-2020 Tier IV, Inc. All rights reserved.
+// Copyright 2015 TIER IV, Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -303,6 +303,20 @@ bool API::attachDetectionSensor(const std::string & entity_name)
 {
   return attachDetectionSensor(helper::constructDetectionSensorConfiguration(
     entity_name, getParameter<std::string>("architecture_type", "awf/universe"), 0.1, 300, false));
+}
+
+bool API::attachOccupancyGridSensor(
+  const simulation_api_schema::OccupancyGridSensorConfiguration & sensor_configuration)
+{
+  if (configuration.standalone_mode) {
+    return true;
+  } else {
+    simulation_api_schema::AttachOccupancyGridSensorRequest req;
+    simulation_api_schema::AttachOccupancyGridSensorResponse res;
+    *req.mutable_configuration() = sensor_configuration;
+    zeromq_client_.call(req, res);
+    return res.result().success();
+  }
 }
 
 bool API::attachLidarSensor(const simulation_api_schema::LidarConfiguration & lidar_configuration)