+++ simulation/traffic_simulator/src/metrics/metrics_manager.cpp
@@ -1,4 +1,4 @@
-// Copyright 2015-2021 TierIV.inc. All rights reserved.
+// Copyright 2015 TIER IV.inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.